fix(PPDSC-2163): address reviews
