+++ lib/codemod/src/cli.js
@@ -18,6 +18,8 @@ const TRANSFORMS = {
     'Unordered List has now a default marker, removes the prop passing the same icon now set as default.',
   'update-list-item-marker-ul-value':
     'Unordered List has now a default marker, the script passes "listItemMarker" with a value of "null" to UnorderedList components originally not passing any marker.',
+  'enum-to-union':
+    'Replace enum type with union type and remove the imports of enums',
 };
 
 function expandFilePathsIfNeeded(filesBeforeExpansion) {
