--- lib/codemod/src/transforms/__tests__/enums-to-union/test-fixture/button-actual.js
@@ -1,3 +1,0 @@
-import {Button, ButtonSize} from 'newskit';
-
-const Component = () => <Button size={ButtonSize.Small}>Button</Button>;
