--- lib/codemod/src/transforms/__tests__/enums-to-union/test-fixture/button-expected.js
@@ -1,3 +1,0 @@
-import {Button} from 'newskit';
-
-const Component = () => <Button size="small">Button</Button>;
