--- lib/codemod/src/transforms/__tests__/enums-to-union/test-fixture/menu-expected.js
@@ -1,9 +1,0 @@
-import {Menu, MenuItem} from 'newskit';
-
-const Component = () => (
-  <Menu vertical align="start" size="small">
-    <MenuItem href={href}>Menu item knickknackatory 1</MenuItem>
-    <MenuItem href={href}>Menu item knickknackatory 2</MenuItem>
-    <MenuItem href={href}>Menu item knickknackatory 3</MenuItem>
-  </Menu>
-);
