--- lib/codemod/src/transforms/__tests__/enums-to-union/test-fixture/slider-actual.js
@@ -1,13 +1,0 @@
-import {Slider, LabelPosition} from 'newskit';
-
-const Component = () => (
-  <Slider
-    values={[50]}
-    max={100}
-    min={0}
-    minLabel="0%"
-    maxLabel="100%"
-    thumbLabel
-    labelPosition={LabelPosition.After}
-  />
-);
