--- lib/codemod/src/transforms/__tests__/enums-to-union/test-fixture/slider-expected.js
@@ -1,13 +1,0 @@
-import {Slider} from 'newskit';
-
-const Component = () => (
-  <Slider
-    values={[50]}
-    max={100}
-    min={0}
-    minLabel="0%"
-    maxLabel="100%"
-    thumbLabel
-    labelPosition="after"
-  />
-);
