--- lib/codemod/src/transforms/__tests__/enums-to-union/test-fixture/stackChild-actual.js
@@ -1,7 +1,0 @@
-import {StackChild, AlignSelfValues} from 'newskit';
-
-const Component = () => (
-  <StackChild alignSelf={AlignSelfValues.Stretch}>
-    <Divider />
-  </StackChild>
-);
