--- lib/codemod/src/transforms/__tests__/enums-to-union/test-fixture/stackChild-expected.js
@@ -1,7 +1,0 @@
-import {StackChild} from 'newskit';
-
-const Component = () => (
-  <StackChild alignSelf="stretch">
-    <Divider />
-  </StackChild>
-);
