--- lib/codemod/src/transforms/__tests__/enums-to-union/test-fixture/tabs-expected.js
@@ -1,21 +1,0 @@
-import {Tabs, Tab} from 'newskit';
-
-const Component = () => (
-  <Tabs
-    size="medium"
-    align="start"
-    vertical
-    indicatorPosition="start"
-    distribution="equal"
-  >
-    <Tab label="Tab">
-      <LoremIpsum textNumber={1} />
-    </Tab>
-    <Tab label="Tab Two">
-      <LoremIpsum textNumber={1} />
-    </Tab>
-    <Tab label="Tab Three is Long">
-      <LoremIpsum textNumber={1} />
-    </Tab>
-  </Tabs>
-);
