--- lib/codemod/src/transforms/__tests__/enums-to-union/test-fixture/textField-actual.js
@@ -1,9 +1,0 @@
-import {FormInput, FormInputTextField, TextFieldSize} from 'newskit';
-
-const Component = () => (
-  <Form onSubmit={onSubmit}>
-    <FormInput state="valid" name="email-valid">
-      <FormInputTextField size={TextFieldSize.Small} />
-    </FormInput>
-  </Form>
-);
