--- lib/codemod/src/transforms/__tests__/enums-to-union/test-fixture/textField-expected.js
@@ -1,9 +1,0 @@
-import {FormInput, FormInputTextField} from 'newskit';
-
-const Component = () => (
-  <Form onSubmit={onSubmit}>
-    <FormInput state="valid" name="email-valid">
-      <FormInputTextField size="small" />
-    </FormInput>
-  </Form>
-);
