+++ lib/codemod/src/utils/__tests__/to-kebab-case.test.js
@@ -1,4 +1,4 @@
-const {toKebabCase} = require('../utils');
+const {toKebabCase} = require('../to-kebab-case');
 
 describe('to-kebab-case', () => {
   test('toKebabCase', () => {
